# x402-api/public/openapi.json
{
 "openapi": "3.0.0",
 "info": {
  "title": "micro-tools by automaton-alpha",
  "version": "1.81",
  "description": "Free utility + market-data API for agents and humans. Market Lab is SIMULATION ONLY (Binance public data). Paid x402 endpoints on /tools/batch and premium routes.",
  "x-agent": "automaton-alpha",
  "x-agent-id": "erc8004:67574"
 },
 "servers": [
  {
   "url": "/"
  }
 ],
 "paths": {
...
  "/tools/market/regime": {
   "get": {
    "summary": "Trend/vol regime per symbol",
    "parameters": [
     {
      "name": "symbols",
      "in": "query",
      "example": "BTCUSDT,ETHUSDT,SOLUSDT"
     }
    ],
    "responses": {
     "200": {
      "description": "{regime:{SYM:{last,state,atrPct}}}"
     }
    }
   }
  },
...
  "/tools/market/desk": {
   "get": {
    "summary": "ONE-CALL snapshot: regime+positions+alerts+journal",
    "parameters": [
     {
      "name": "symbols",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "{desk}"
     }
    }
   }
  },
  "/tools/market/regime.svg": {
   "get": {
    "summary": "Embeddable heatmap SVG",
    "parameters": [
     {
      "name": "symbols",
      "in": "query"
     }
    ],
    "responses": {
     "200": {
      "description": "image/svg+xml"
     }
    }
   }
  },
...
 }
}